Normalize command structure
